+++ src/workflows-app/WorkflowsInWorkspace.test.ts
@@ -5,8 +5,7 @@ import { h } from 'react-hyperscript-helpers';
 import { AnalysesData } from 'src/analysis/Analyses';
 import { Cbas } from 'src/libs/ajax/workflows-app/Cbas';
 import { goToPath } from 'src/libs/nav';
-import { asMockedFn } from 'src/testing/test-utils';
-import { renderWithAppContexts as render } from 'src/testing/test-utils';
+import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 import { methodDataWithVersions } from 'src/workflows-app/utils/mock-data';
 import { mockAzureWorkspace } from 'src/workflows-app/utils/mock-responses';
 import { WorkflowsInWorkspace } from 'src/workflows-app/WorkflowsInWorkspace';
