+++ src/workflows-app/utils/submission-utils.ts
@@ -2,7 +2,6 @@ import _ from 'lodash/fp';
 import { div } from 'react-hyperscript-helpers';
 import { icon } from 'src/components/icons';
 import { statusType as jobStatusType } from 'src/components/job-common';
-import { Ajax } from 'src/libs/ajax';
 import colors from 'src/libs/colors';
 import * as Utils from 'src/libs/utils';
 import { differenceFromDatesInSeconds, differenceFromNowInSeconds, maybeParseJSON } from 'src/libs/utils';
@@ -44,34 +43,6 @@ export type OutputTableData = {
   outputType: OutputType;
   taskName: string;
   variable: string;
-};
-
-export type WorkflowMetadata = {
-  actualWorkflowLanguage: string;
-  actualWorkflowLanguageVersion: string;
-  calls: {};
-  end: string;
-  id: string;
-  inputs: {}[];
-  labels: {};
-  outputs: {}[];
-  start: string;
-  status: string;
-  submission: string;
-  submittedFiles: {};
-  workflowCallback: {};
-  workflowLog: string;
-  workflowName: string;
-  workflowProcessingEvents: {}[];
-  workflowRoot: string;
-};
-
-export type MetadataOptions = {
-  cromwellProxyUrl: string;
-  excludeKeys: string[];
-  includeKeys: string[];
-  signal: AbortSignal;
-  workflowId: string;
 };
 
 const iconSize = 24;
@@ -607,8 +578,3 @@ export const getOutputTableData = (
     _.orderBy([({ [sort.field]: field }) => _.lowerCase(field)], [sort.direction])
   )(configuredOutputDefinition);
 };
-
-export const fetchMetadata = async (options: MetadataOptions): Promise<WorkflowMetadata> =>
-  Ajax(options.signal)
-    .CromwellApp.workflows(options.workflowId)
-    .metadata(options.cromwellProxyUrl, { includeKey: options.includeKeys, excludeKey: options.excludeKeys });
